+++ clients/rust/src/models/octo_print_event.rs
@@ -14,7 +14,7 @@ pub struct OctoPrintEvent {
     #[serde(rename = "id")]
     pub id: i32,
     #[serde(rename = "ts", skip_serializing_if = "Option::is_none")]
-    pub ts: Option<i32>,
+    pub ts: Option<f32>,
     #[serde(rename = "event_type")]
     pub event_type: crate::models::OctoPrintEventEventTypeEnum,
     #[serde(rename = "octoprint_environment")]
