+++ clients/rust/src/models/remote_command_event.rs
@@ -14,7 +14,7 @@ pub struct RemoteCommandEvent {
     #[serde(rename = "id")]
     pub id: i32,
     #[serde(rename = "ts", skip_serializing_if = "Option::is_none")]
-    pub ts: Option<i32>,
+    pub ts: Option<f32>,
     #[serde(rename = "event_type")]
     pub event_type: crate::models::RemoteCommandEventEventTypeEnum,
     #[serde(rename = "octoprint_environment")]
